src/requirements.txt
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
langchain==0.2.12
langchain-community==0.2.11
langchain-core~=0.2
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.99
python-dotenv==1.0.1
shortuuid==1.0.13
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
gunicorn==23.0.0
icecream==2.1.3
django-pwa==2.0.1
psycopg2-binary==2.9.10
django-phonenumber-field==8.0.0
phonenumberslite==8.13.55
whitenoise==6.9.0
django-recaptcha==4.0.0
Faker==37.0.0
weasyprint==64.1
